sales table now obeys rls
