+++ src/app/dashboard/sales/page.tsx
@@ -9,9 +9,7 @@ import Link from "next/link";
 import Button from "~/components/ui/Button";
 
 export default function Page() {
-  const { isLoading, data } = api.sales.allSales.useQuery({
-    organizationEmail: ""
-  });
+  const { isLoading, data } = api.sales.allSales.useQuery();
 
   return (
     <Layout>
@@ -20,7 +18,7 @@ export default function Page() {
         <div className="flex w-[90vw] items-center justify-between  pt-[40px] md:w-[1000px]">
           <h3 className="text-2xl font-medium md:text-3xl ">Your Company </h3>
           <div className="flex items-center gap-2">
-            <Link href="/sales/new">
+            <Link href="/dashboard/sales/new">
               <Button>Make a Sale</Button>
             </Link>
           </div>
@@ -30,7 +28,7 @@ export default function Page() {
             bigTitle="Looks like you haven't added any sales yet"
             smallTitle="When you make sales, they will appear here"
             c2a="Make a Sale"
-            c2aUrl="/sales/new"
+            c2aUrl="/dashboard/sales/new"
           />
         )}
         {data === null && (
